pyproject.toml
[project]
name = "iamcompact-vetting"
version = "0.0.2"
description = "Vetting checks for model outputs in the HORIZON EUROPE project IAM COMPACT"
authors = [
    { name = "Jan Ivar Korsbakken", email = "janko@cicero.oslo.no" }
]
readme = "README.md"
requires-python = ">= 3.11"
dependencies = [
    "pyam-iamc>=2.1",
    "pathways-ensemble-analysis @ git+https://gitlab.com/climateanalytics/pathways-ensemble-analysis.git@v1.1.0",
    # Dependency required by p-e-a (but not listed among its requirements)
    "pandas-indexing>=0.5.1",
    "openpyxl>=3.1.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ipython>=8.20.0",
    "pytest>=8.0.0",
    "ipykernel>=6.29.0",
    "dvc[ssh]>=3.48.4",
    "pytest-cov>=5.0.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["iamcompact_vetting"]
